Added AI.
Added OpenAI functionality that requires your OpenAI API Key to talk to the companion.

Added the ability to change the voice recording keybind to suit your needs.

Added the ability to customize the text (size, colour, position) that the AI response displays.

AI replies with text that is displayed on top of the container that Yazuki is in.

Changed the font as some letters weren't included, swapped out for a nicer font.